Structure clean-up
Moved everything to it's own package.

activities - contains all the activity classes with business logic;
adapters - contains all the adapters used in the project;
fragments - Is going to contain all the new fragments;
interfaces - Is going to contain all the new interfaces that are not services;
models - Response classes for now. They are basically models at this point;
network - contains the network logic;
services - contains all the services;
utils - is going to contain the helper classes that will be created.

This structure can be found on: https://guides.codepath.com/android/Organizing-your-Source-Files